project-01-sre-platform/platform-api/pyproject.toml
[project]
name = "platform-api"
version = "0.1.0"
description = "Internal developer platform API — Project 01 of the sre-platform portfolio"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Moulima Das" }]

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "kubernetes>=31.0.0",
  "hvac>=2.3.0",
  "httpx>=0.27.0",
  "prometheus-client>=0.21.0",
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-exporter-otlp>=1.27.0",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-instrumentation-httpx>=0.48b0",
  "google-genai>=0.7.0",
  "ulid-py>=1.1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Strip the leading 'src/' so modules import flatly (e.g. `import config`
# rather than `from src.config import ...`). The Dockerfile sets
# PYTHONPATH=/app/src to match this convention at runtime.
sources = ["src"]
# Every subdirectory under src/ that has __init__.py is a package.
packages = ["src/k8s", "src/vault", "src/slo", "src/receipts", "src/llm", "src/routes"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "B", "UP", "RUF", "SIM", "PL"]
ignore = [
  "E501",
  "PLR0913",
  "PLR2004",
  "PLW0603",  # intentional module-level singleton pattern
  "PLC0415",  # lazy imports for optional dependencies
  "RUF002",   # dash flavour is stylistic
  "RUF003",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR2004", "S101"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
  "ignore::DeprecationWarning",
]
